refactor: LLM provider를 의존성으로 주입받는 형태로 개선
