+++ src/ai/ai.interface.ts
@@ -1,0 +1,5 @@
+import { CallAnalysisResult } from './types';
+
+export interface AiAnalysisProvider {
+  analyze(transcript: string): Promise<CallAnalysisResult>;
+}
